requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
pytest==8.0.0
python-dotenv==1.0.1
selenium==4.17.2
